% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_fix.R
\docType{data}
\name{AUTO_FIX_SYSTEM_PROMPT}
\alias{AUTO_FIX_SYSTEM_PROMPT}
\title{System Prompt for Auto-Fix}
\format{
An object of class \code{character} of length 1.
}
\usage{
AUTO_FIX_SYSTEM_PROMPT
}
\description{
System Prompt for Auto-Fix
}
\keyword{internal}
