% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/artifacts.R
\name{create_artifact_dir}
\alias{create_artifact_dir}
\title{Artifact Tools for File Persistence}
\usage{
create_artifact_dir(base_dir = NULL)
}
\description{
Artifact Tools for File Persistence
}
\keyword{internal}
