% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\name{extract_theme_for_frontend}
\alias{extract_theme_for_frontend}
\title{Extract Theme for Frontend}
\usage{
extract_theme_for_frontend(theme, plot_dims)
}
\description{
Extracts theme with structured units and pixel values.
}
\keyword{internal}
