% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\name{ggplot_to_frontend_object}
\alias{ggplot_to_frontend_object}
\title{Convert ggplot to Frontend-Friendly Object}
\usage{
ggplot_to_frontend_object(plot, include_data, plot_dims)
}
\description{
Internal function for structured conversion.
}
\keyword{internal}
