% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_engine.R
\name{knitr_engine}
\alias{knitr_engine}
\title{Knitr Engine for AI}
\description{
Implements a custom knitr engine \code{{ai}} that allows using LLMs to generate
and execute R code within RMarkdown/Quarto documents.
}
