% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\docType{data}
\name{KNOWN_AESTHETICS}
\alias{KNOWN_AESTHETICS}
\title{Known Aesthetic Types}
\format{
An object of class \code{list} of length 24.
}
\usage{
KNOWN_AESTHETICS
}
\description{
Registry of known aesthetics with their expected types.
}
\keyword{internal}
