% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\docType{data}
\name{KNOWN_POSITIONS}
\alias{KNOWN_POSITIONS}
\title{Known Position Types with Parameters}
\format{
An object of class \code{list} of length 7.
}
\usage{
KNOWN_POSITIONS
}
\description{
Known Position Types with Parameters
}
\keyword{internal}
