% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\name{sanitize_theme_element}
\alias{sanitize_theme_element}
\title{Sanitize Theme Element for Frontend}
\usage{
sanitize_theme_element(elem, plot_dims)
}
\description{
Sanitize Theme Element for Frontend
}
\keyword{internal}
