% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\docType{data}
\name{THEME_ELEMENT_TYPES}
\alias{THEME_ELEMENT_TYPES}
\title{Theme Element Types}
\format{
An object of class \code{list} of length 6.
}
\usage{
THEME_ELEMENT_TYPES
}
\description{
Registry of theme element types and their properties.
}
\keyword{internal}
