% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\docType{data}
\name{THEME_HIERARCHY}
\alias{THEME_HIERARCHY}
\title{Theme Component Hierarchy}
\format{
An object of class \code{list} of length 54.
}
\usage{
THEME_HIERARCHY
}
\description{
Defines the hierarchical structure of theme components.
}
\keyword{internal}
