% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\name{z_aes_mapping}
\alias{z_aes_mapping}
\title{Aesthetic Mapping Schema}
\usage{
z_aes_mapping(known_only = FALSE)
}
\arguments{
\item{known_only}{If TRUE, only include known aesthetics in schema.}
}
\value{
A z_object schema.
}
\description{
Schema for ggplot2 aesthetic mappings (aes).
}
