% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\name{z_geom_layer}
\alias{z_geom_layer}
\title{Build Geom-Specific Layer Schema}
\usage{
z_geom_layer(geom_name)
}
\arguments{
\item{geom_name}{Name of the geom.}
}
\value{
A z_object schema tailored to the geom.
}
\description{
Creates a precise schema for a specific geom type.
}
