% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\name{z_theme}
\alias{z_theme}
\title{Theme Schema}
\usage{
z_theme(flat = TRUE)
}
\arguments{
\item{flat}{If TRUE, returns flat structure. If FALSE, returns hierarchical.}
}
\value{
A z_object schema.
}
\description{
Schema for theme settings with full hierarchy support.
}
