% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_schema.R
\name{z_theme_element}
\alias{z_theme_element}
\title{Create Element Schema}
\usage{
z_theme_element(element_type)
}
\arguments{
\item{element_type}{One of the THEME_ELEMENT_TYPES names.}
}
\value{
A z_object schema.
}
\description{
Creates a schema for a specific theme element type.
}
\keyword{internal}
